+++ app/ThirdParty/ThirdPartySupport.php
@@ -122,6 +122,7 @@ class ThirdPartySupport {
 	 */
 	private function get_third_party_support(): array {
 		$list = array(
+			'ExternalFilesInMediaLibrary\ThirdParty\AdvancedMediaOffloader',
 			'ExternalFilesInMediaLibrary\ThirdParty\BlockEditor',
 			'ExternalFilesInMediaLibrary\ThirdParty\CatFolders',
 			'ExternalFilesInMediaLibrary\ThirdParty\Downloadlist',